+++ history.md
@@ -1,4 +1,26 @@
 # 版本信息
+## 3.12.0 (2020-09-27)
+* feat: 更新 `yyl-seed-webpack@2.5.13` 兼容 当 `yconfig.localserver.entry` 时 `writeDisk` 自动设为 true
+* feat: 更新 `yyl-server@0.3.17` 兼容 `seed-other` 出现的问题
+* feat: 更新 `yyl-hander@0.11.4` 允许 `yyl.config.js` 只设置 `localserver` `proxy` 和 只设置 `proxy` 情况
+* feat: 允许 `yyl server start --port false` 禁用本地server
+* feat: 新增 `yyl-seed-other@0.1.1`
+* feat: 兼容 `fec`, `feb`, `yys` 构建方案
+* feat: 新增 `yyl-seed-gulp-other@0.1.1`
+* feat: 更新 `yyl-seed-gulp-requirejs@4.7.3`
+* feat: 当配置 `config.yarn: true` 时， 会自动删除 项目根目录中的 `package-lock.json` 文件
+* fix: 修复 `yyl-seed-gulp-requirejs` 项目中 html 文件 `__url(path/to/file)` 语法糖失效问题
+* fix: 修复 当 `config.commit.hostname === '//www.yy.com/pc'` 这种带有 `pathname` 的 配置时， `localserver` 映射不正确问题
+
+## 3.11.7 (2020-07-22)
+* feat: 更新 `yyl-seed-gulp-requirejs@4.7.1`
+* feat: 更新 `yyl-seed-webpack@2.5.9`
+* feat: 更换 seed 包中 `node-sass` -> `sass`
+## 3.11.6 (2020-07-22)
+* feat: 更新 `yyl-seed-gulp-requirejs@4.7.0`
+* feat: 更新 `yyl-seed-webpack@2.5.8`
+* feat: 兼容 node@14
+
 ## 3.11.5 (2020-07-08)
 * feat: 更新 `yyl-server@0.3.10` 修复 ios 13 不能正常代理https 问题
 * feat: 更新 `yyl-os@0.11.2` 修复打开带[&] url 会报错问题